modules refactor, push_thread methods
